feat: #194 - ダッシュボードデータ取得最適化
