+++ src/components/dashboard/interview-list-skeleton.tsx
@@ -1,0 +1,38 @@
+import { Skeleton } from "@/components/ui/skeleton";
+
+/**
+ * ダッシュボード面接一覧のスケルトン表示
+ * Suspense の fallback として使用
+ */
+export function InterviewListSkeleton() {
+  return (
+    <div className="mt-6">
+      <div className="grid gap-4 sm:grid-cols-2 lg:grid-cols-3">
+        {Array.from({ length: 6 }).map((_, i) => (
+          <div key={i} className="rounded-xl border p-4 space-y-3">
+            {/* 企業名 */}
+            <div className="flex items-center gap-2">
+              <Skeleton className="h-4 w-4 rounded" />
+              <Skeleton className="h-5 w-32" />
+            </div>
+            {/* バッジ */}
+            <div className="flex gap-1.5">
+              <Skeleton className="h-5 w-16 rounded-full" />
+              <Skeleton className="h-5 w-12 rounded-full" />
+            </div>
+            {/* 日付 */}
+            <Skeleton className="h-4 w-28" />
+            {/* スコアバー */}
+            <div className="space-y-1.5">
+              <div className="flex justify-between">
+                <Skeleton className="h-3 w-14" />
+                <Skeleton className="h-3 w-10" />
+              </div>
+              <Skeleton className="h-1.5 w-full rounded-full" />
+            </div>
+          </div>
+        ))}
+      </div>
+    </div>
+  );
+}
